[Caio] Fixed bug in create_user_service

+++ models/claim.py
@@ -4,7 +4,7 @@ class Claim(db.Model):
     __tablename__ = 'claims'
     __table_args__ = {'schema': 'shipay_db'}  
 
-    id = db.Column(db.BigInteger, primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     description = db.Column(db.String, nullable=False)
     active = db.Column(db.Boolean, default=True, nullable=False)
